requirements.txt

flask-sqlALCHEMY
FLASK-LOGIN
FLASK-RESTFUL
gunicorn
psycopg2
flask-bootstrap
libsass
FLASK-MIGRATE